+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/TelegramBot.kt
@@ -77,7 +77,6 @@ class TelegramBot(
         bot = this,
         classManager = config.classManager,
         inputListener = config.inputListener,
-        rateLimiter = config.rateLimiter,
     )
 
     internal var httpClient = getConfiguredHttpClient()