+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/core/EnvConfigLoader.kt
@@ -1,6 +1,7 @@
 package eu.vendeli.tgbot.core
 
 import ch.qos.logback.classic.Level
+import com.fasterxml.jackson.module.kotlin.isKotlinClass
 import eu.vendeli.tgbot.interfaces.ChatData
 import eu.vendeli.tgbot.interfaces.ClassManager
 import eu.vendeli.tgbot.interfaces.ConfigLoader
@@ -9,6 +10,7 @@ import eu.vendeli.tgbot.interfaces.RateLimitMechanism
 import eu.vendeli.tgbot.interfaces.UserData
 import eu.vendeli.tgbot.types.internal.HttpLogLevel
 import eu.vendeli.tgbot.types.internal.configuration.BotConfiguration
+import kotlin.reflect.full.primaryConstructor
 
 object EnvConfigLoader : ConfigLoader {
     private const val PREFIX = "TGBOT_"
@@ -21,7 +23,12 @@ object EnvConfigLoader : ConfigLoader {
     override val commandsPackage: String? = getVal("COMMANDS_PACKAGE")
 
     private fun getVal(value: String) = envVars[PREFIX + value]
-    private fun String.init() = Class.forName(this).constructors.first().newInstance()
+    private fun String.init(): Any? = Class.forName(this).run {
+        takeIf { isKotlinClass() }?.kotlin.let {
+            it?.objectInstance ?: it?.primaryConstructor?.call()
+        } ?: constructors.firstOrNull()?.newInstance()
+    }
+
     private fun apply2Config(block: BotConfiguration.() -> Unit) = BotConfiguration().also(block)
 
     @Suppress("CyclomaticComplexMethod")
@@ -30,7 +37,7 @@ object EnvConfigLoader : ConfigLoader {
         getVal("API_HOST")?.also { apiHost = it }
         getVal("INPUT_LISTENER")?.also { inputListener = it.init() as InputListener }
         getVal("CLASS_MANAGER")?.also { classManager = it.init() as ClassManager }
-        getVal("RATE_LIMITER")?.also { rateLimiter = it.init() as RateLimitMechanism }
+        getVal("RATE_LIMITER")?.also { rateLimiter.mechanism = it.init() as RateLimitMechanism }
         // httpClient
         getVal("HTTPC_RQ_TIMEOUT_MILLIS")?.toLongOrNull()?.also { httpClient.requestTimeoutMillis = it }
         getVal("HTTPC_C_TIMEOUT_MILLIS")?.toLongOrNull()?.also { httpClient.connectTimeoutMillis = it }
@@ -41,8 +48,8 @@ object EnvConfigLoader : ConfigLoader {
         getVal("LOG_BOT_LVL")?.also { logging.botLogLevel = Level.toLevel(it) }
         getVal("LOG_HTTP_LVL")?.also { logging.httpLogLevel = HttpLogLevel.valueOf(it) }
         // rateLimits
-        getVal("RATES_PERIOD")?.toLongOrNull()?.also { rateLimits.period = it }
-        getVal("RATES_RATE")?.toLongOrNull()?.also { rateLimits.rate = it }
+        getVal("RATES_PERIOD")?.toLongOrNull()?.also { rateLimiter.limits.period = it }
+        getVal("RATES_RATE")?.toLongOrNull()?.also { rateLimiter.limits.rate = it }
         // context
         getVal("CTX_USER_DATA")?.also { context.userData = it.init() as UserData }
         getVal("CTX_CHAT_DATA")?.also { context.chatData = it.init() as ChatData }