+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/core/TokenBucketLimiterImpl.kt
@@ -25,7 +25,7 @@ internal data class BucketState(
 /**
  * Default implementation of query limitation via [Token bucket](https://en.wikipedia.org/wiki/Token_bucket) algorithm.
  */
-open class TokenBucketLimiterImpl : RateLimitMechanism {
+class TokenBucketLimiterImpl : RateLimitMechanism {
     private val state: ConcurrentHashMap<String, AtomicReference<BucketState>> = ConcurrentHashMap()
     private val instant: Instant
         get() {