+++ telegram-bot/src/test/kotlin/eu/vendeli/InputListenerTest.kt
@@ -2,7 +2,9 @@ package eu.vendeli
 
 import BotTestContext
 import eu.vendeli.tgbot.core.InputListenerMapImpl
+import eu.vendeli.tgbot.types.User
 import io.kotest.matchers.booleans.shouldBeTrue
+import io.kotest.matchers.nulls.shouldBeNull
 import io.kotest.matchers.shouldBe
 
 class InputListenerTest : BotTestContext() {
@@ -13,6 +15,14 @@ class InputListenerTest : BotTestContext() {
         mapImpl.set(1, "test")
         mapImpl.get(1) shouldBe "test"
         mapImpl.del(1)
+        mapImpl.get(1).shouldBeNull()
+
+        val user = User(1, false, "Test")
+        mapImpl.set(user) { "test2" }
+        mapImpl[user] shouldBe "test2"
+
+        mapImpl[user] = "test3"
+        mapImpl[user] shouldBe "test3"
     }
 
     @Test
@@ -20,5 +30,10 @@ class InputListenerTest : BotTestContext() {
         mapImpl.setAsync(1, "test").await().shouldBeTrue()
         mapImpl.getAsync(1).await() shouldBe "test"
         mapImpl.delAsync(1).await().shouldBeTrue()
+        mapImpl.getAsync(1).await().shouldBeNull()
+
+        val user = User(1, false, "Test")
+        mapImpl.setAsync(user) { "test2" }.await().shouldBeTrue()
+        mapImpl.getAsync(user.id).await() shouldBe "test2"
     }
 }