+++ telegram-bot/src/test/kotlin/eu/vendeli/RateLimitingTest.kt
@@ -2,27 +2,21 @@ package eu.vendeli
 
 import BotTestContext
 import ch.qos.logback.classic.Level
-import eu.vendeli.tgbot.TelegramBot
-import eu.vendeli.tgbot.core.TokenBucketLimiterImpl
 import eu.vendeli.tgbot.types.internal.configuration.RateLimits
 import io.kotest.core.spec.IsolationMode
 import io.kotest.matchers.shouldBe
 import java.util.concurrent.atomic.AtomicInteger
 
-class RateLimitingTest : BotTestContext(false, true) {
+class RateLimitingTest : BotTestContext(mockHttp = true) {
     override fun isolationMode(): IsolationMode = IsolationMode.InstancePerTest
 
-    @BeforeAll
-    fun prepareBot() {
-        bot = TelegramBot("not necessary") {
-            rateLimiter = TokenBucketLimiterImpl()
-            logging {
-                botLogLevel = Level.INFO
-            }
-            rateLimits {
-                period = 10000
-                rate = 5
-            }
+    @BeforeEach
+    fun prepareBot() = bot.config.run {
+        rateLimiter {
+            limits = RateLimits(10000, 5)
+        }
+        logging {
+            botLogLevel = Level.DEBUG
         }
     }